Organise java files into packages
