+++ src/main/java/dash/Ui.java
@@ -1,3 +1,5 @@
+package dash;
+
 import java.util.ArrayList;
 import java.util.NoSuchElementException;
 import java.util.Scanner;
