+++ src/main/java/dash/Utils.java
@@ -1,3 +1,5 @@
+package dash;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.List;
