+++ src/main/java/dash/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package dash.command;
+
+import dash.exception.ExitException;
+import dash.task.TaskList;
+import dash.Ui;
+
 public class ByeCommand implements Command {
     @Override
     public void execute(TaskList taskList, Ui ui) {
