+++ src/main/java/dash/command/Command.java
@@ -1,4 +1,7 @@
-import java.util.ArrayList;
+package dash.command;
+
+import dash.task.TaskList;
+import dash.Ui;
 
 public interface Command {
     public void execute(TaskList taskList, Ui ui);
