+++ src/main/java/dash/command/CommandEnum.java
@@ -1,3 +1,5 @@
+package dash.command;
+
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
