+++ src/main/java/dash/command/DeleteAllCommand.java
@@ -1,3 +1,8 @@
+package dash.command;
+
+import dash.task.TaskList;
+import dash.Ui;
+
 public class DeleteAllCommand implements Command {
     public void execute(TaskList taskList, Ui ui) {
         taskList.clear();
