+++ src/main/java/dash/exception/ExitException.java
@@ -1,3 +1,5 @@
+package dash.exception;
+
 public class ExitException extends RuntimeException {
     public ExitException() {
         super("Exit/bye command received.");
