+++ src/main/java/dash/exception/UnknownCommandException.java
@@ -1,3 +1,5 @@
+package dash.exception;
+
 public class UnknownCommandException extends RuntimeException {
     public UnknownCommandException(String message) {
         super(message);
