+++ src/main/java/dash/task/Deadline.java
@@ -1,3 +1,5 @@
+package dash.task;
+
 import java.time.LocalDate;
 
 public class Deadline extends Task {
