+++ src/main/java/dash/task/Event.java
@@ -1,3 +1,5 @@
+package dash.task;
+
 import java.time.LocalDate;
 
 public class Event extends Task {
