+++ src/main/java/dash/task/Task.java
@@ -1,3 +1,5 @@
+package dash.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
