+++ src/main/java/dash/task/Todo.java
@@ -1,3 +1,5 @@
+package dash.task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
